Move operations into their own directory
